Move Layout to main directory
